Move logger spy setup to ag-charts-community test helper to avoid circular dependency
The logger spy setup needs to be in ag-charts-community test helper instead of
ag-charts-test because ag-charts-test cannot import ag-charts-core without
creating a circular dependency. The community test helper already has access
to both Logger (from ag-charts-core) and the base setupMockConsole (from
ag-charts-test), making it the right place for this logic.
