+++ libraries/ag-charts-test/src/console/mock-console.ts
@@ -12,7 +12,6 @@ const ALL_METHODS = [...NORMAL_METHODS, ...ERROR_METHODS];
 // debug option to disable the message suppression.
 export function setupMockConsole(debugShowOutput?: boolean, { includeAllLevels = false } = {}) {
     const mocks = new Map<string, [jest.Mock, Function]>();
-    const loggerSpies = new Map<string, jest.SpyInstance>();
     const consoleMethods = [...ERROR_METHODS];
     if (includeAllLevels) {
         consoleMethods.push(...NORMAL_METHODS);
@@ -27,68 +26,27 @@ export function setupMockConsole(debugShowOutput?: boolean, { includeAllLevels =
     }
 
     beforeAll(() => {
-        // Mock console methods first
         for (const method of consoleMethods) {
             const fn = console[method];
             const mock = createConsoleMock(console[method]);
             console[method] = mock;
             mocks.set(method, [mock, fn]);
         }
-
-        // Spy on logger exports - they use bound console methods, so we need to spy on them
-        // and make them delegate to the console mocks
-        // Use dynamic require to avoid circular dependency issues at build time
-        try {
-            // eslint-disable-next-line @typescript-eslint/no-var-requires
-            const loggerModule = require('ag-charts-core/src/globals/logger');
-            const loggerMethods = ['warn', 'error', 'warnOnce', 'errorOnce'] as const;
-            
-            for (const method of loggerMethods) {
-                if (loggerModule[method]) {
-                    const targetConsoleMethod = method === 'warnOnce' ? 'warn' : method === 'errorOnce' ? 'error' : method;
-                    const consoleMock = mocks.get(targetConsoleMethod)?.[0];
-                    if (consoleMock) {
-                        const spy = jest.spyOn(loggerModule, method).mockImplementation((...args: any[]) => {
-                            // Delegate to the console mock
-                            // The bound functions prepend 'AG Charts - ' as the first arg, so we pass all args
-                            return consoleMock(...args);
-                        });
-                        loggerSpies.set(method, spy);
-                    }
-                }
-            }
-        } catch (e) {
-            // Logger module might not be available in all test contexts, ignore silently
-            // This can happen if ag-charts-core is not a dependency of the test package
-        }
     });
 
     afterEach(() => {
         try {
-            // Check both console mocks and logger spies for errors
             for (const method of ERROR_METHODS) {
                 expect(mocks.get(method)?.[0]).not.toHaveBeenCalled();
             }
-            const warnSpy = loggerSpies.get('warn');
-            const errorSpy = loggerSpies.get('error');
-            if (warnSpy) expect(warnSpy).not.toHaveBeenCalled();
-            if (errorSpy) expect(errorSpy).not.toHaveBeenCalled();
         } finally {
             for (const method of consoleMethods) {
                 mocks.get(method)?.[0].mockClear();
-            }
-            for (const spy of loggerSpies.values()) {
-                spy.mockClear();
             }
         }
     });
 
     afterAll(() => {
-        // Restore logger spies
-        for (const spy of loggerSpies.values()) {
-            spy.mockRestore();
-        }
-        // Restore console methods
         for (const method of consoleMethods) {
             console[method] = mocks.get(method)?.[1] as any;
         }
